services/crm_service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
prometheus-fastapi-instrumentator>=6.1.0
# CRM Adapters
hubspot-api-client>=9.0.0
simple-salesforce>=1.12.0
# Testing
pytest>=7.0.0
httpx>=0.24.0
